+++ examples/sessions/session.py
@@ -1,0 +1,10 @@
+from aim import Session
+
+sess = Session()
+
+sess.set_params({
+    'foo': 'bar',
+})
+
+for i in range(10):
+    sess.track(i, name='val')